Moved 'animations.js' to /config
