+++ config/animations.js
