+++ hooks/animations/interaction/useMagnetic.js
@@ -1,6 +1,6 @@
 import { useRef, useState, useCallback, useEffect } from "react";
 import gsap from "gsap";
-import { MAGNETIC as getMagnetic } from "@/constants/animations";
+import { MAGNETIC as getMagnetic } from "@/config/animations";
 import { useHoverCapable } from "@/hooks/utils/useHoverCapable";
 import { useReducedMotion } from "@/context/ReducedMotionContext";
