+++ hooks/animations/onload/useHeroTitleOnloadAnimations.js
@@ -1,7 +1,7 @@
 import { useLayoutEffect, useState } from "react";
 import { gsap } from "gsap/dist/gsap";
 import { useGSAP } from "@gsap/react";
-import { HERO as getHero } from "@/constants/animations";
+import { HERO as getHero } from "@/config/animations";
 import { useStartedLines } from "@/context/PreloaderContext";
 import { useWindowDimensions } from "@/context/DimensionsContext";
 import { useTransition } from "@/context/TransitionContext";
