+++ hooks/animations/scroll/useWorkScrollAnimations.js
@@ -2,7 +2,7 @@ import { useRef, useState, useLayoutEffect, useMemo } from "react";
 import { gsap } from "gsap/dist/gsap";
 import { useGSAP } from "@gsap/react";
 import { ScrollTrigger } from "gsap/dist/ScrollTrigger";
-import { WORK as getWork } from "@/constants/animations";
+import { WORK as getWork } from "@/config/animations";
 import styles from "@/components/Home/Work/Work.module.css";
 import { useTransition } from "@/context/TransitionContext";
 import { useReducedMotion } from "@/context/ReducedMotionContext";
