+++ hooks/animations/transition/useContactTransitions.js
@@ -1,7 +1,7 @@
 import { useRef } from "react";
 import { gsap } from "gsap/dist/gsap";
 import { useGSAP } from "@gsap/react";
-import { CONTACT as getContact } from "@/constants/animations";
+import { CONTACT as getContact } from "@/config/animations";
 
 export const useContactTransition = (isOpen) => {
   const CONTACT = getContact();
